+++ packages/datepicker/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @heathmont/moon-datepicker
 
+## 9.1.2
+
+### Patch Changes
+
+- Template
+- Updated dependencies [undefined]
+  - @heathmont/moon-components@9.1.2
+  - @heathmont/moon-core@9.1.2
+  - @heathmont/moon-icons@9.1.2
+  - @heathmont/moon-themes@9.1.2
+  - @heathmont/moon-utils@9.1.2
+
 ## 9.1.1
 
 ### Patch Changes